@@ -77,88 +75,120 @@
     "js-test-jquery": "cross-env JQUERY=true npm run js-test-karma",
     "lint": "npm-run-all --aggregate-output --continue-on-error --parallel js-lint css-lint lockfile-lint",
     "docs": "npm-run-all docs-build docs-lint",
-    "docs-build": "hugo --cleanDestinationDir",
+    "docs-build": "npm run astro-build",
     "docs-compile": "npm run docs-build",
-    "docs-vnu": "node build/vnu-jar.js",
-    "docs-lint": "npm run docs-vnu",
-    "docs-serve": "hugo server --port 9001 --disableFastRender",
+    "docs-vnu": "node build/vnu-jar.mjs",
+    "docs-lint": "npm-run-all docs-prettier-check docs-vnu",
+    "docs-prettier-check": "prettier --config site/.prettierrc.json -c --cache site",
+    "docs-prettier-format": "prettier --config site/.prettierrc.json --write --cache site",
+    "docs-serve": "npm run astro-dev",
     "docs-serve-only": "npx sirv-cli _site --port 9001",
     "lockfile-lint": "lockfile-lint --allowed-hosts npm --allowed-schemes https: --empty-hostname false --type npm --path package-lock.json",
-    "update-deps": "ncu -u -x globby,karma-browserstack-launcher,karma-rollup-preprocessor && echo Manually update site/assets/js/vendor",
+    "update-deps": "ncu -u -x eslint,eslint-config-xo,eslint-plugin-unicorn,karma-browserstack-launcher,karma-rollup-preprocessor,sass",
     "release": "npm-run-all dist release-sri docs-build release-zip*",
-    "release-sri": "node build/generate-sri.js",
-    "release-version": "node build/change-version.js",
-    "release-zip": "cross-env-shell \"rm -rf bootstrap-$npm_package_version-dist && cp -r dist/ bootstrap-$npm_package_version-dist && zip -r9 bootstrap-$npm_package_version-dist.zip bootstrap-$npm_package_version-dist && rm -rf bootstrap-$npm_package_version-dist\"",
-    "release-zip-examples": "node build/zip-examples.js",
+    "release-sri": "node build/generate-sri.mjs",
+    "release-version": "node build/change-version.mjs",
+    "release-zip": "cross-env-shell \"rm -rf bootstrap-$npm_package_version-dist bootstrap-$npm_package_version-dist.zip && cp -r dist/ bootstrap-$npm_package_version-dist && zip -qr9 bootstrap-$npm_package_version-dist.zip bootstrap-$npm_package_version-dist && rm -rf bootstrap-$npm_package_version-dist\"",
+    "release-zip-examples": "node build/zip-examples.mjs",
     "dist": "npm-run-all --aggregate-output --parallel css js",
     "test": "npm-run-all lint dist js-test docs-build docs-lint",
-    "netlify": "cross-env-shell HUGO_BASEURL=$DEPLOY_PRIME_URL npm-run-all dist release-sri docs-build",
+    "netlify": "npm-run-all dist release-sri astro-build",
     "watch": "npm-run-all --parallel watch-*",
     "watch-css-main": "nodemon --watch scss/ --ext scss --exec \"npm-run-all css-lint css-compile css-prefix\"",
     "watch-css-dist": "nodemon --watch dist/css/ --ext css --ignore \"dist/css/*.rtl.*\" --exec \"npm run css-rtl\"",
-    "watch-css-docs": "nodemon --watch site/assets/scss/ --ext scss --exec \"npm run css-lint\"",
+    "watch-css-docs": "nodemon --watch site/src/scss/ --ext scss --exec \"npm run css-lint\"",
+    "watch-css-test": "nodemon --watch scss/ --ext scss,js --exec \"npm run css-test\"",
     "watch-js-main": "nodemon --watch js/src/ --ext js --exec \"npm-run-all js-lint js-compile\"",
-    "watch-js-docs": "nodemon --watch site/assets/js/ --ext js --exec \"npm run js-lint\""
+    "watch-js-docs": "nodemon --watch site/src/assets/ --ext js --exec \"npm run js-lint\"",
+    "astro-dev": "astro dev --root site --port 9001",
+    "astro-build": "astro build --root site && rm -rf _site && cp -r site/dist _site",
+    "astro-preview": "astro preview --root site --port 9001"
   },
   "peerDependencies": {
-    "@popperjs/core": "^2.11.5"
+    "@popperjs/core": "^2.11.8"
   },
   "devDependencies": {
-    "@babel/cli": "^7.18.9",
-    "@babel/core": "^7.18.9",
-    "@babel/preset-env": "^7.18.9",
-    "@popperjs/core": "^2.11.5",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-replace": "^4.0.0",
-    "autoprefixer": "^10.4.7",
-    "bundlewatch": "^0.3.3",
-    "clean-css-cli": "^5.6.1",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.20.0",
-    "eslint-config-xo": "^0.41.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-markdown": "^3.0.0",
-    "eslint-plugin-unicorn": "^42.0.0",
-    "find-unused-sass-variables": "^4.0.4",
-    "globby": "^11.1.0",
+    "@astrojs/check": "^0.9.4",
+    "@astrojs/markdown-remark": "^6.3.6",
+    "@astrojs/mdx": "^4.3.4",
+    "@astrojs/prism": "^3.3.0",
+    "@astrojs/sitemap": "^3.5.0",
+    "@babel/cli": "^7.28.3",
+    "@babel/core": "^7.28.3",
+    "@babel/preset-env": "^7.28.3",
+    "@docsearch/js": "^3.9.0",
+    "@popperjs/core": "^2.11.8",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^28.0.6",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-replace": "^6.0.2",
+    "@stackblitz/sdk": "^1.11.0",
+    "@types/js-yaml": "^4.0.9",
+    "@types/mime": "^4.0.0",
+    "@types/prismjs": "^1.26.5",
+    "astro": "^5.13.2",
+    "astro-auto-import": "^0.4.4",
+    "autoprefixer": "^10.4.21",
+    "bundlewatch": "^0.4.1",
+    "clean-css-cli": "^5.6.3",
+    "clipboard": "^2.0.11",
+    "cross-env": "^10.0.0",
+    "eslint": "8.57.1",
+    "eslint-config-xo": "0.45.0",
+    "eslint-plugin-html": "^8.1.3",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-markdown": "^5.1.0",
+    "eslint-plugin-unicorn": "56.0.1",
+    "find-unused-sass-variables": "^6.1.0",
+    "github-slugger": "^2.0.0",
+    "globby": "^14.1.0",
     "hammer-simulator": "0.0.1",
-    "hugo-bin": "^0.89.0",
-    "ip": "^2.0.0",
-    "jquery": "^3.6.0",
-    "karma": "^6.4.0",
+    "htmlparser2": "^10.0.0",
+    "image-size": "^2.0.2",
+    "ip": "^2.0.1",
+    "jasmine": "^5.9.0",
+    "jquery": "^3.7.1",
+    "js-yaml": "^4.1.0",
+    "karma": "^6.4.4",
     "karma-browserstack-launcher": "1.4.0",
-    "karma-chrome-launcher": "^3.1.1",
+    "karma-chrome-launcher": "^3.2.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-detect-browsers": "^2.3.3",
-    "karma-firefox-launcher": "^2.1.2",
+    "karma-firefox-launcher": "^2.1.3",
     "karma-jasmine": "^5.1.0",
-    "karma-jasmine-html-reporter": "^2.0.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "karma-rollup-preprocessor": "7.0.7",
-    "lockfile-lint": "^4.7.6",
-    "nodemon": "^2.0.19",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.14",
-    "postcss-cli": "^10.0.0",
-    "rollup": "^2.77.0",
-    "rollup-plugin-istanbul": "^3.0.0",
-    "rtlcss": "^3.5.0",
-    "sass": "^1.53.0",
-    "shelljs": "^0.8.5",
-    "stylelint": "^14.9.1",
-    "stylelint-config-twbs-bootstrap": "^5.0.0",
-    "terser": "^5.14.2",
-    "vnu-jar": "21.10.12"
+    "lockfile-lint": "^4.14.1",
+    "mime": "^4.0.7",
+    "nodemon": "^3.1.10",
+    "npm-run-all2": "^8.0.4",
+    "postcss": "^8.5.6",
+    "postcss-cli": "^11.0.1",
+    "prettier": "^3.6.2",
+    "prettier-plugin-astro": "^0.14.1",
+    "rehype-autolink-headings": "^7.1.0",
+    "remark": "^15.0.1",
+    "remark-html": "^16.0.1",
+    "rollup": "^4.46.3",
+    "rollup-plugin-istanbul": "^5.0.0",
+    "rtlcss": "^4.3.0",
+    "sass": "1.78.0",
+    "sass-true": "^9.0.0",
+    "shelljs": "^0.10.0",
+    "stylelint": "^16.23.1",
+    "stylelint-config-twbs-bootstrap": "^16.1.0",
+    "terser": "^5.43.1",
+    "unist-util-visit": "^5.0.0",
+    "vnu-jar": "24.10.17",
+    "zod": "^4.0.17"
   },
   "files": [
     "dist/{css,js}/*.{css,js,map}",
     "js/{src,dist}/**/*.{js,map}",
-    "scss/**/*.scss"
+    "js/index.{esm,umd}.js",
+    "scss/**/*.scss",
+    "!scss/tests/**"
   ],
-  "hugo-bin": {
-    "buildTags": "extended"
-  },
   "jspm": {
     "registry": "npm",
     "main": "js/bootstrap",
