@@ -5,8 +5,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+
 @Repository
-public interface MenuRepository extends JpaRepository<Menu,Long> {
-    List<Menu> findByRestaurant_Id(Long id);
-    void deleteByRestaurant_Id(Long id);
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+    // Get all menus for a specific restaurant
+    List<Menu> findByRestaurantId(Long restaurantId);
+
+	void deleteByRestaurant_Id(Long res_id);
+
+	List<Menu> findByRestaurant_Id(Long id);
 }
