@@ -1,3 +1,12 @@
+//package com.fooddeliveryapp.foodappbackend.repository;
+//
+//import com.fooddeliveryapp.foodappbackend.entity.Restaurant;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface RestaurantRepository extends JpaRepository<Restaurant,Long> {
+//}
 package com.fooddeliveryapp.foodappbackend.repository;
 
 import com.fooddeliveryapp.foodappbackend.entity.Restaurant;
@@ -5,5 +14,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RestaurantRepository extends JpaRepository<Restaurant,Long> {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
