@@ -1,8 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { UserLoginComponent } from './user-login/user-login.component';
+import { UserRegisterComponent } from './user-register/user-register.component';
+import { HomeComponent } from './home/home.component';
 
 const routes: Routes = [
-  { path:'', redirectTo: '/home', pathMatch: 'full'},
+  { path: '', redirectTo: '/login', pathMatch: 'full' },
+  { path: 'login', component: UserLoginComponent },
+  { path: 'signup', component: UserRegisterComponent },
+  { path: 'home', component: HomeComponent },
 ];
 
 @NgModule({
