@@ -1,14 +1,7 @@
 export class UserDto {
-
-    public id: number = 0;
-
-    public username: string ="";
-
-    public password: string ="";
-
-    public userType: string ="";
-
-    public email: string ="";
-
-    public confirmpassword: string ="";
+  firstName!: string;
+  lastName!: string;
+  email!: string;
+  password!: string;
+  username: any;
 }
