@@ -1,4 +1,4 @@
 export class UserLoginRequest {
-    public username: string ="";
-    public password: string ="";
+  email!: string;
+  password!: string;
 }
