@@ -1,20 +1,16 @@
-import { Injectable } from "@angular/core";
-import { Observable } from "rxjs";
-import { UserLoginRequest } from "../dto/UserLoginRequest";
 import { HttpClient } from "@angular/common/http";
+import { Injectable } from "@angular/core";
+import { UserLoginRequest } from "../dto/UserLoginRequest";
+import { Observable } from "rxjs";
 
-
-@Injectable({
-    providedIn: 'root'
-})
+@Injectable({ providedIn: 'root' })
 export class UserLoginService {
 
-    private url = 'http://localhost:9099/api/v1.0/login';
+  private url = 'http://localhost:8080/api/v1.0/login';   // ✅ Correct URL
 
-    constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
-    validateLoginAndGetConfirmedUser(userLoginRequest: UserLoginRequest): Observable<any>{
-        return this.http.post<any>(this.url, userLoginRequest);
-    }
-
+  validateLoginAndGetConfirmedUser(req: UserLoginRequest): Observable<any> {
+    return this.http.post<any>(this.url, req);
+  }
 }
